api/requirements.txt
Django
psycopg2 
djangorestframework 
django-cors-headers
djangorestframework-simplejwt
PyJWT
djoser